email-ai-agent/pyproject.toml
[project]
name = "email-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "autogen>=0.10.4",
    "litellm>=1.81.6",
    "pyyaml>=6.0.3",
    "resend>=2.21.0",
]
